refactor: move tests inside tests directory
