+++ src/view/calendar/tests/custom-event-free-busy-status.test.tsx
@@ -8,8 +8,8 @@ import React from 'react';
 import { screen } from '@testing-library/react';
 import { useTheme } from 'styled-components';
 
-import { CustomEventFreeBusyStatus } from './custom-event-free-busy-status';
-import { EVENT_DISPLAY_STATUS } from '../../constants/api';
+import { EVENT_DISPLAY_STATUS } from '../../../constants/api';
+import { CustomEventFreeBusyStatus } from '../custom-event-free-busy-status';
 import { setupHook, setupTest } from '@test-setup';
 
 describe('CustomEventFreeBusyStatus', () => {
