+++ src/view/calendar/tests/custom-event.test.tsx
@@ -8,10 +8,10 @@ import React from 'react';
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 import { act, screen } from '@testing-library/react';
 
-import { MemoCustomEvent } from './custom-event';
-import { reducers } from '../../store/redux';
-import { useAppStatusStore } from '../../store/zustand/store';
-import mockedData from '../../test/generators';
+import { reducers } from '../../../store/redux';
+import { useAppStatusStore } from '../../../store/zustand/store';
+import mockedData from '../../../test/generators';
+import { MemoCustomEvent } from '../custom-event';
 import { setupTest } from '@test-setup';
 
 jest.setTimeout(10000);
